Make all io connetion become module
